# goye-web/backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"SignupResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {},
					"token": {}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.Exclude_keyofUser.id__": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email_address": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"createAt": {},
					"updatedAt": {}
				},
				"required": [
					"first_name",
					"last_name",
					"email_address",
					"password",
					"country",
					"state",
					"phone_number",
					"role",
					"level"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.id_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Partial_User_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email_address": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"createAt": {},
					"updatedAt": {}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
		"/user/signup": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignupResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_User.id_"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/sendOtp": {
			"post": {
				"operationId": "SendOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/verify-otp": {
			"post": {
				"operationId": "VerifyOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"email": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"email": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"email",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"sessionToken": {
										"type": "string"
									},
									"otp": {
										"type": "string"
									}
								},
								"required": [
									"sessionToken",
									"otp"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/upload-profile-picture": {
			"post": {
				"operationId": "UploadPicture",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"mimeType": {
										"type": "string"
									},
									"fileName": {
										"type": "string"
									},
									"file": {
										"type": "string"
									}
								},
								"required": [
									"mimeType",
									"fileName",
									"file"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/get-user/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/update-user/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignupResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_User_"
							}
						}
					}
				}
			}
		},
		"/user/delete-user/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"properties": {
												"user_pic": {
													"type": "string"
												},
												"lastActive": {
													"type": "string",
													"format": "date-time"
												},
												"isOnline": {
													"type": "boolean"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"completed": {
													"type": "string"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"level": {
													"type": "string"
												},
												"role": {
													"type": "string"
												},
												"phone_number": {
													"type": "string"
												},
												"state": {
													"type": "string"
												},
												"country": {
													"type": "string"
												},
												"password": {
													"type": "string"
												},
												"email_address": {
													"type": "string"
												},
												"last_name": {
													"type": "string"
												},
												"first_name": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"user_pic",
												"lastActive",
												"isOnline",
												"createdAt",
												"completed",
												"updatedAt",
												"level",
												"role",
												"phone_number",
												"state",
												"country",
												"password",
												"email_address",
												"last_name",
												"first_name",
												"id"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/user/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"user": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"user": {
													"properties": {
														"user_pic": {
															"type": "string"
														},
														"lastActive": {
															"type": "string",
															"format": "date-time"
														},
														"isOnline": {
															"type": "boolean"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"completed": {
															"type": "string"
														},
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"level": {
															"type": "string"
														},
														"role": {
															"type": "string"
														},
														"phone_number": {
															"type": "string"
														},
														"state": {
															"type": "string"
														},
														"country": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"email_address": {
															"type": "string"
														},
														"last_name": {
															"type": "string"
														},
														"first_name": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"user_pic",
														"lastActive",
														"isOnline",
														"createdAt",
														"completed",
														"updatedAt",
														"level",
														"role",
														"phone_number",
														"state",
														"country",
														"password",
														"email_address",
														"last_name",
														"first_name",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"user",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/user/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"link": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"link",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/user/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User control APIs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}